Refactor PerformTaskNode to move task completion method into their own classes.
Add classes ConstructFromStructureTemplate, ConstructFromBuildingGenerator and PlantTaskCompletion.
